feat(types): first-class u8 byte width
bytes indexing always returned u8, but the surrounding surface spoke i32:
the std/io shim family was typed i32, printing and assertions had no u8
arms, and the byte-focused test suites asserted through widening casts.
u8 becomes a first-class width — literal range checks, display and
f-string support, testing assertions, and the JIT symbol rows — the
std/io bytes shims are retyped to u8 with every consumer migrated, and
the byte suites assert byte-width values directly. The checker's
no-implicit-widening rule is untouched throughout.

+++ hew-types/tests/declarative_string_bytes_ffi_differential.rs
@@ -90,9 +90,9 @@ fn bytes_methods_resolve_through_std_io_extern_symbols() {
     let source = r"
         fn exercise(buf: bytes, other: bytes) {
             buf.push(65);
-            let _: i32 = buf.pop();
+            let _: u8 = buf.pop();
             let _: i64 = buf.len();
-            let _: i32 = buf.get(0);
+            let _: u8 = buf.get(0);
             buf.set(0, 66);
             let _: bool = buf.is_empty();
             buf.clear();
